+++ backend-cotisation/src/main/java/org/sid/cotisation/entities/RegionProjection.java
@@ -4,7 +4,7 @@ import org.springframework.data.rest.core.config.Projection;
 
 @Projection(name = "Rg", types = Region.class)
 public interface RegionProjection {
-    public int getId();
+    public Long getId();
     public String getNom();
     public Pays getPays();
 }
